Refactor integration tests with base class IntegrationTestBase
