+++ src/integrationTest/java/uk/gov/hmcts/reform/preapi/services/AuditServiceIT.java
@@ -10,20 +10,19 @@ import uk.gov.hmcts.reform.preapi.enums.AuditAction;
 import uk.gov.hmcts.reform.preapi.enums.AuditLogSource;
 import uk.gov.hmcts.reform.preapi.enums.CourtType;
 import uk.gov.hmcts.reform.preapi.util.HelperFactory;
+import uk.gov.hmcts.reform.preapi.utils.IntegrationTestBase;
 
-@SpringBootTest(classes = Application.class)
-public class AuditServiceIT {
-
-    @Autowired
-    CourtService courtService;
+public class AuditServiceIT extends IntegrationTestBase {
 
     @Autowired
-    AuditService auditService;
+    private CourtService courtService;
+
+    @Autowired
+    private AuditService auditService;
 
     @Transactional
     @Test
     public void testInternalAudit() {
-
         var court = HelperFactory.createCreateCourtDTO(CourtType.CROWN, "Foo Court", "1234");
         courtService.upsert(court);
