+++ src/integrationTest/java/uk/gov/hmcts/reform/preapi/services/CaseServiceIT.java
@@ -21,6 +21,7 @@ import uk.gov.hmcts.reform.preapi.exception.RecordingNotDeletedException;
 import uk.gov.hmcts.reform.preapi.repositories.RecordingRepository;
 import uk.gov.hmcts.reform.preapi.security.authentication.UserAuthentication;
 import uk.gov.hmcts.reform.preapi.util.HelperFactory;
+import uk.gov.hmcts.reform.preapi.utils.IntegrationTestBase;
 
 import java.sql.Timestamp;
 import java.time.Instant;
@@ -32,30 +33,12 @@ import java.util.stream.Collectors;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-@SpringBootTest(classes = Application.class)
-public class CaseServiceIT {
-    @Autowired
-    private EntityManager entityManager;
-
+public class CaseServiceIT extends IntegrationTestBase {
     @Autowired
     private CaseService caseService;
+
     @Autowired
     private RecordingRepository recordingRepository;
-
-    public static void mockAdminUser() {
-        var mockAuth = mock(UserAuthentication.class);
-        when(mockAuth.isAdmin()).thenReturn(true);
-        when(mockAuth.isAppUser()).thenReturn(true);
-        SecurityContextHolder.getContext().setAuthentication(mockAuth);
-    }
-
-    public static void mockNonAdminUser() {
-        var mockAuth = mock(UserAuthentication.class);
-        when(mockAuth.isAdmin()).thenReturn(false);
-        when(mockAuth.isAppUser()).thenReturn(true);
-        when(mockAuth.isPortalUser()).thenReturn(false);
-        SecurityContextHolder.getContext().setAuthentication(mockAuth);
-    }
 
     @Transactional
     @Test
