+++ src/integrationTest/java/uk/gov/hmcts/reform/preapi/services/UserServiceIT.java
@@ -22,6 +22,7 @@ import uk.gov.hmcts.reform.preapi.entities.User;
 import uk.gov.hmcts.reform.preapi.enums.AccessType;
 import uk.gov.hmcts.reform.preapi.enums.CourtType;
 import uk.gov.hmcts.reform.preapi.security.authentication.UserAuthentication;
+import uk.gov.hmcts.reform.preapi.utils.IntegrationTestBase;
 
 import java.sql.Timestamp;
 import java.time.Instant;
@@ -31,8 +32,7 @@ import java.util.UUID;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-@SpringBootTest(classes = Application.class)
-public class UserServiceIT {
+public class UserServiceIT extends IntegrationTestBase {
     private static User userEntity;
     private static User portalUserEntity;
     private static User appUserEntity;
@@ -42,9 +42,6 @@ public class UserServiceIT {
     private static PortalAccess portalAccessEntity2;
     private static Court court;
     private static Role role;
-
-    @Autowired
-    private EntityManager entityManager;
 
     @Autowired
     private UserService userService;
@@ -113,26 +110,6 @@ public class UserServiceIT {
         portalAccessEntity2.setId(UUID.randomUUID());
         portalAccessEntity2.setUser(portalUserEntity);
         entityManager.persist(portalAccessEntity2);
-    }
-
-    @AfterEach
-    void tearDown() {
-        entityManager.clear();
-        entityManager.flush();
-    }
-
-    public static void mockAdminUser() {
-        var mockAuth = mock(UserAuthentication.class);
-        when(mockAuth.isAdmin()).thenReturn(true);
-        when(mockAuth.isAppUser()).thenReturn(true);
-        SecurityContextHolder.getContext().setAuthentication(mockAuth);
-    }
-
-    public static void mockNonAdminUser() {
-        var mockAuth = mock(UserAuthentication.class);
-        when(mockAuth.isAdmin()).thenReturn(false);
-        when(mockAuth.isAppUser()).thenReturn(true);
-        SecurityContextHolder.getContext().setAuthentication(mockAuth);
     }
 
     @Transactional
