+++ src/integrationTest/java/uk/gov/hmcts/reform/preapi/utils/IntegrationTestBase.java
@@ -1,0 +1,42 @@
+package uk.gov.hmcts.reform.preapi.utils;
+
+import jakarta.persistence.EntityManager;
+import org.junit.jupiter.api.AfterEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.core.context.SecurityContextHolder;
+import uk.gov.hmcts.reform.preapi.Application;
+import uk.gov.hmcts.reform.preapi.security.authentication.UserAuthentication;
+
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+@SpringBootTest(classes = Application.class)
+public abstract class IntegrationTestBase {
+
+    @Autowired
+    protected EntityManager entityManager;
+
+    protected static UserAuthentication mockAdminUser() {
+        var mockAuth = mock(UserAuthentication.class);
+        when(mockAuth.isAdmin()).thenReturn(true);
+        when(mockAuth.isAppUser()).thenReturn(true);
+        SecurityContextHolder.getContext().setAuthentication(mockAuth);
+        return mockAuth;
+    }
+
+    protected static UserAuthentication mockNonAdminUser() {
+        var mockAuth = mock(UserAuthentication.class);
+        when(mockAuth.isAdmin()).thenReturn(false);
+        when(mockAuth.isAppUser()).thenReturn(true);
+        when(mockAuth.isPortalUser()).thenReturn(false);
+        SecurityContextHolder.getContext().setAuthentication(mockAuth);
+        return mockAuth;
+    }
+
+    @AfterEach
+    void tearDown() {
+        entityManager.clear();
+        entityManager.flush();
+    }
+}
